+++ packages/react-components/src/organisms/ResearchOutputFormSharingCard.tsx
@@ -1,8 +1,6 @@
 import {
-  DecisionOption,
   ResearchOutputPostRequest,
   ResearchOutputResponse,
-  ResearchOutputSharingStatus,
   ResearchOutputType,
   ResearchTagResponse,
   ValidationErrorResponse,
@@ -16,10 +14,8 @@ import { MultiSelectOptionsType } from '../atoms/MultiSelect';
 import { GlobeIcon } from '../icons';
 import {
   FormCard,
-  LabeledDateField,
   LabeledDropdown,
   LabeledMultiSelect,
-  LabeledRadioButtonGroup,
   LabeledTextArea,
   LabeledTextEditor,
   LabeledTextField,
@@ -32,9 +28,9 @@ type ResearchOutputFormSharingCardProps = Pick<
   | 'link'
   | 'title'
   | 'descriptionMD'
+  | 'layImpactStatement'
   | 'shortDescription'
   | 'changelog'
-  | 'sharingStatus'
   | 'subtype'
 > & {
   isFormSubmitted: boolean;
@@ -54,25 +50,17 @@ type ResearchOutputFormSharingCardProps = Pick<
     ComponentPropsWithRef<typeof LabeledMultiSelect>['onChange']
   >;
   displayChangelog?: boolean;
-  researchOutputData?: ResearchOutputResponse;
   documentType?: ResearchOutputResponse['documentType'];
-  isCreatingOutputRoute?: boolean;
   type?: ResearchOutputType | '';
   onChangeLink?: (newValue: string) => void;
   onChangeTitle?: (newValue: string) => void;
   onChangeDescription?: (newValue: string) => void;
   onChangeShortDescription?: (newValue: string) => void;
+  onChangeLayImpactStatement?: (newValue: string) => void;
   onChangeChangelog?: (newValue: string) => void;
   onChangeType?: (newValue: ResearchOutputType | '') => void;
   onChangeSubtype?: (newValue: string | '') => void;
-  onChangeAsapFunded?: (newValue: DecisionOption) => void;
-  onChangeUsedInPublication?: (newValue: DecisionOption) => void;
-  onChangeSharingStatus?: (newValue: ResearchOutputSharingStatus) => void;
-  onChangePublishDate?: (newValue: Date | undefined) => void;
   isSaving: boolean;
-  asapFunded: DecisionOption;
-  usedInPublication: DecisionOption;
-  publishDate?: Date;
   researchTags: ResearchTagResponse[];
   serverValidationErrors?: ValidationErrorResponse['data'];
   clearServerValidationError?: (instancePath: string) => void;
@@ -80,14 +68,6 @@ type ResearchOutputFormSharingCardProps = Pick<
   urlRequired?: boolean;
   typeOptions: ResearchOutputType[];
   getShortDescriptionFromDescription: (description: string) => Promise<string>;
-  isImportedFromManuscript?: boolean;
-};
-
-export const getPublishDateValidationMessage = (e: ValidityState): string => {
-  if (e.badInput) {
-    return 'Date published should be complete or removed';
-  }
-  return 'Publish date cannot be greater than today';
 };
 
 const ResearchOutputFormSharingCard: React.FC<
@@ -97,14 +77,13 @@ const ResearchOutputFormSharingCard: React.FC<
   isCreatingNewVersion,
   getImpactSuggestions,
   impact,
+  layImpactStatement,
   onChangeImpact,
   getCategorySuggestions = noop,
   categories,
   onChangeCategories,
   displayChangelog = false,
-  researchOutputData,
   documentType,
-  isCreatingOutputRoute,
   isSaving,
   link,
   title,
@@ -114,15 +93,10 @@ const ResearchOutputFormSharingCard: React.FC<
   type,
   typeOptions,
   subtype,
-  asapFunded,
-  usedInPublication,
-  sharingStatus,
-  publishDate,
   researchTags,
   serverValidationErrors = [],
   typeDescription,
   urlRequired,
-  isImportedFromManuscript,
   getShortDescriptionFromDescription,
   clearServerValidationError = noop,
   onChangeDescription = noop,
@@ -132,10 +106,7 @@ const ResearchOutputFormSharingCard: React.FC<
   onChangeTitle = noop,
   onChangeType = noop,
   onChangeSubtype = noop,
-  onChangeAsapFunded = noop,
-  onChangeUsedInPublication = noop,
-  onChangeSharingStatus = noop,
-  onChangePublishDate = noop,
+  onChangeLayImpactStatement = noop,
 }) => {
   const [urlValidationMessage, setUrlValidationMessage] = useState<string>();
   const [titleValidationMessage, setTitleValidationMessage] =
@@ -316,63 +287,6 @@ const ResearchOutputFormSharingCard: React.FC<
           placeholder="Select subtype"
         />
       )}
-
-      {documentType === 'Article' && (
-        <LabeledDropdown
-          required
-          getValidationMessage={() => 'Please choose an impact.'}
-          title="Impact"
-          subtitle="(required)"
-          description="Select the option that best describes the impact of this manuscript on the PD field."
-          options={impactOptions}
-          onChange={(e) => {
-            const impactOption = impactOptions.find(
-              (option) => option.value === e,
-            );
-            onChangeImpact(
-              impactOption as MultiSelectOptionsType &
-                OptionsType<MultiSelectOptionsType>,
-            );
-          }}
-          onBlur={validateImpact}
-          customValidationMessage={impactValidationMessage}
-          value={impact?.value ?? ''}
-          enabled={!isSaving && !isCreatingNewVersion}
-          noOptionsMessage={(option) =>
-            `Sorry, no impacts match ${option.inputValue}`
-          }
-          placeholder="Choose an impact"
-        />
-      )}
-      {documentType === 'Article' && (
-        <LabeledMultiSelect
-          required
-          getValidationMessage={(validationState) =>
-            validationState.valueMissing
-              ? 'Please add at least one category.'
-              : 'You can select up to two categories only.'
-          }
-          title="Category"
-          description="Select up to two options that best describe the scientific category of this manuscript."
-          subtitle="(required)"
-          enabled={!isSaving && !isCreatingNewVersion}
-          placeholder="Start typing..."
-          loadOptions={getCategorySuggestions}
-          onChange={(newValues) => {
-            onChangeCategories(
-              newValues as MultiSelectOptionsType &
-                OptionsType<MultiSelectOptionsType>,
-            );
-            validateCategories(newValues);
-          }}
-          customValidationMessage={categoryValidationMessage}
-          values={categories as OptionsType<MultiSelectOptionsType>}
-          noOptionsMessage={({ inputValue }) =>
-            `Sorry, no category options match ${inputValue}`
-          }
-        />
-      )}
-
       <LabeledTextEditor
         title="Description"
         subtitle="(required)"
@@ -407,6 +321,77 @@ const ResearchOutputFormSharingCard: React.FC<
         }
         customValidationMessage={shortDescriptionValidationMessage}
       />
+
+      {documentType === 'Article' && (
+        <>
+          <LabeledDropdown
+            required
+            getValidationMessage={() => 'Please choose an impact.'}
+            title="Impact"
+            subtitle="(required)"
+            description="Select the option that best describes the impact of this manuscript on the PD field."
+            options={impactOptions}
+            onChange={(e) => {
+              const impactOption = impactOptions.find(
+                (option) => option.value === e,
+              );
+              onChangeImpact(
+                impactOption as MultiSelectOptionsType &
+                  OptionsType<MultiSelectOptionsType>,
+              );
+            }}
+            onBlur={validateImpact}
+            customValidationMessage={impactValidationMessage}
+            value={impact?.value ?? ''}
+            enabled={!isSaving && !isCreatingNewVersion}
+            noOptionsMessage={(option) =>
+              `Sorry, no impacts match ${option.inputValue}`
+            }
+            placeholder="Choose an impact"
+          />
+          <LabeledTextArea
+            value={layImpactStatement ?? ''}
+            onChange={onChangeLayImpactStatement}
+            enabled={!isSaving}
+            required
+            maxLength={100}
+            title="Lay Impact Statement"
+            subtitle="(required)"
+            tip={
+              <span>
+                Explain in plain language why this work matters and how it may
+                impact research, patients, or the wider community.
+              </span>
+            }
+          />
+          <LabeledMultiSelect
+            required
+            getValidationMessage={(validationState) =>
+              validationState.valueMissing
+                ? 'Please add at least one category.'
+                : 'You can select up to two categories only.'
+            }
+            title="Category"
+            description="Select up to two options that best describe the scientific category of this manuscript."
+            subtitle="(required)"
+            enabled={!isSaving && !isCreatingNewVersion}
+            placeholder="Start typing..."
+            loadOptions={getCategorySuggestions}
+            onChange={(newValues) => {
+              onChangeCategories(
+                newValues as MultiSelectOptionsType &
+                  OptionsType<MultiSelectOptionsType>,
+              );
+              validateCategories(newValues);
+            }}
+            customValidationMessage={categoryValidationMessage}
+            values={categories as OptionsType<MultiSelectOptionsType>}
+            noOptionsMessage={({ inputValue }) =>
+              `Sorry, no category options match ${inputValue}`
+            }
+          />
+        </>
+      )}
       {displayChangelog && (
         <LabeledTextArea
           title="Changelog"
@@ -426,75 +411,6 @@ const ResearchOutputFormSharingCard: React.FC<
           enabled={!isSaving}
         />
       )}
-      <LabeledRadioButtonGroup
-        title="Has this output been funded by ASAP"
-        subtitle="(required)"
-        options={[
-          { value: 'Yes', label: 'Yes' },
-          { value: 'No', label: 'No' },
-          { value: 'Not Sure', label: 'Not Sure' },
-        ]}
-        value={asapFunded}
-        onChange={onChangeAsapFunded}
-      />
-
-      <LabeledRadioButtonGroup
-        title="Has this output been used in a publication"
-        subtitle="(required)"
-        options={[
-          { value: 'Yes', label: 'Yes' },
-          {
-            value: 'No',
-            label: 'No',
-            disabled:
-              isCreatingOutputRoute &&
-              documentType === 'Article' &&
-              researchOutputData?.usedInPublication === undefined,
-          },
-          {
-            value: 'Not Sure',
-            label: 'Not Sure',
-            disabled:
-              isCreatingOutputRoute &&
-              documentType === 'Article' &&
-              researchOutputData?.usedInPublication === undefined,
-          },
-        ]}
-        value={usedInPublication}
-        onChange={onChangeUsedInPublication}
-      />
-
-      <LabeledRadioButtonGroup
-        title="Sharing status"
-        subtitle="(required)"
-        options={[
-          {
-            value: 'Network Only',
-            label: 'CRN Only',
-            disabled:
-              (documentType === 'Article' &&
-                researchOutputData?.sharingStatus === undefined) ||
-              isImportedFromManuscript,
-          },
-          { value: 'Public', label: 'Public' },
-        ]}
-        value={sharingStatus}
-        onChange={onChangeSharingStatus}
-      />
-
-      {sharingStatus === 'Public' ? (
-        <LabeledDateField
-          title={'Date Published'}
-          subtitle={'(optional)'}
-          description={
-            'This should be the date your output was shared publicly on its repository.'
-          }
-          onChange={onChangePublishDate}
-          value={publishDate}
-          max={new Date()}
-          getValidationMessage={(e) => getPublishDateValidationMessage(e)}
-        />
-      ) : null}
     </FormCard>
   );
 };
